app/requirements.txt
tensorflow==1.12.0
numpy
pillow
flask
matplotlib
